fix : CI handler moved to pycm_ci.py #237
